[BE] 마커용 가게 전체 조회 api 추가

+++ src/main/java/com/kinopio/eatgo/store/dto/StoreDetailResponseDto.java
@@ -23,8 +23,8 @@ public class StoreDetailResponseDto {
 	private Integer storeId;
 	private String storeName;
 	private String address;
-	private Integer positionX;
-	private Integer positionY;
+	private Double positionX;
+	private Double positionY;
 	private Integer isOpen;
 	private String thumbnail;
 	private Integer createdType;
